refactor(data): change files names to english
