+++ src/clients/chile-data-fetcher.js
@@ -1,13 +1,13 @@
 import axios from 'axios';
 import camelcaseKeys from 'camelcase-keys';
 import env from '../environment';
-import regiones from '../data/places/regiones';
-import comunas from '../data/places/comunas';
+import regions from '../data/places/regions';
+import communes from '../data/places/communes';
 
 async function getChileData() {
   const response = await axios.get(env.chileDataJsonUrl);
   const { data } = response;
-  return camelcaseKeys(data, { deep: true, exclude: [...regiones, ...comunas] });
+  return camelcaseKeys(data, { deep: true, exclude: [...regions, ...communes] });
 }
 
 export default getChileData;
